[240930] 3. 게시글조회     - src/app/boards/detail/page.tsx 파일의 경로를 다이나믹하게 변경하기 위해 src/app/boards/[boardId]/page.tsx 경로로 변경해 주세요.         - 해당 페이지에 접속하였을 때 주소 부분의 boardId 값을 꺼내오고, 꺼낸 id의 게시물을 GRAPHQL-API(fetchBoard)로 조회해 보세요.         - 조회된 게시물의 작성자, 제목, 내용을 알맞은 위치에 보여주세요.         - 데이터를 받아오는데 시간이 걸리는 이유로, 받아오기 전의 존재하지 않는 데이터를 그리는 과정에서 에러가 발생하게 됩니다. 이 문제를 해결해 주세요.
