+++ 02-main_camp/Day23/src/app/boards/detail/[boardId]/styles.module.css
